Respect the /quiet command line option when converting content. (#8689)
Prevents a "Skipping XYZ..." message for each asset which is _not_
converted during the content build process.

On projects with a lot of content, this creates a lot of noise. I left
the normal output message enabled since the conversion can be slow, and
it gives the user indication of what is happening. The "Skipping XYZ..."
messages just tend to drown out everything else.